+++ .secrets/.gitignore
@@ -1,0 +1,6 @@
+# IMPORTANT! This folder is hidden from git - if you need to store config files or other secrets,
+# make sure those are never staged for commit into your git repo. You can store them here or another
+# secure location.
+
+*
+!.gitignore
