+++ tap_clockify/client.py
@@ -1,31 +1,33 @@
+"""REST client handling, including ClockifyStream base class."""
+
 import requests
-import singer
+from pathlib import Path
+from typing import Any, Dict, Optional, Union, List, Iterable
 
-LOGGER = singer.get_logger()  # noqa
+from singer_sdk.helpers.jsonpath import extract_jsonpath
+from singer_sdk.streams import RESTStream
 
 
-class ClockifyClient:
+SCHEMAS_DIR = Path(__file__).parent / Path("./schemas")
 
-    MAX_TRIES = 5
 
-    def __init__(self, config):
-        self.config = config
+class ClockifyStream(RESTStream):
+    """Clockify stream class."""
 
-    def make_request(self, url, method, params=None, body=None):
-        LOGGER.info("Making %s request to %s (%s)", method, url, params)
+    records_jsonpath = "$[*]"
+    next_page_token_jsonpath = "$.next_page"
 
-        response = requests.request(
-            method,
-            url,
-            headers={
-                "x-api-key": self.config["api_key"],
-                "Content-Type": "application/json",
-            },
-            params=params,
-            json=body,
-        )
+    @property
+    def url_base(self):
+        return f'https://api.clockify.me/api/v1/workspaces/{self.config["workspace"]}'
 
-        if response.status_code != 200:
-            raise RuntimeError(response.text)
-
-        return response.json()
+    @property
+    def http_headers(self) -> dict:
+        """Return the http headers needed."""
+        headers = {
+            "Content-Type": "application/json",
+            "x-api-key": self.config["api_key"]
+        }
+        if "user_agent" in self.config:
+            headers["User-Agent"] = self.config.get("user_agent")
+        return headers
