+++ tox.ini
@@ -1,0 +1,26 @@
+# This file can be used to customize dox tests as well as other test frameworks like flake8 and mypy
+
+[tox]
+envlist = py38
+; envlist = py36, py38, py39
+isolated_build = true
+
+[testenv]
+whitelist_externals = poetry
+
+commands =
+    poetry install -v
+    poetry run pytest
+    poetry run black --check tap_clockify/
+    poetry run flake8 tap_clockify
+    poetry run pydocstyle tap_clockify
+    poetry run mypy tap_clockify --exclude='tap_clockify/tests'
+
+
+[flake8]
+ignore = W503
+max-line-length = 88
+max-complexity = 10
+
+[pydocstyle]
+ignore = D105,D203,D213
